worker/requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aioitertools==0.13.0
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
arq==0.28.0
asyncpg==0.31.0
attrs==26.1.0
av==17.0.1
bcrypt==4.0.1
boto3==1.40.61
botocore==1.40.61
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==46.0.4
ctranslate2==4.7.1
DeepFilterLib==0.5.6
DeepFilterNet==0.5.6
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.128.0
faster-whisper==1.2.1
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.4.3
hiredis==3.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.13.0
HyperPyYAML==1.2.3
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
loguru==0.7.3
Mako==1.3.12
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.25.1
orjson==3.11.5
packaging==23.2
passlib==1.7.4
propcache==0.4.1
protobuf>=3.20.2,<6.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
google-generativeai==0.8.3
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
redis==5.3.1
regex==2026.4.4
requests==2.33.1
rich-toolkit==0.17.1
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
ruff==0.15.12
s3transfer==0.14.0
safetensors==0.7.0
scipy==1.17.1
sentencepiece==0.2.1
setuptools==70.2.0
six==1.17.0
soundfile==0.13.1
speechbrain==1.1.0
SQLAlchemy==2.0.46
sqlmodel==0.0.31
starlette==0.50.0
sympy==1.13.1
tokenizers==0.22.2
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
tqdm==4.67.3
transformers==5.7.0
triton==3.2.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
wrapt==1.17.3
yarl==1.23.0
